moved security constants
