+++ segmentstore/server/host/src/test/java/io/pravega/segmentstore/server/host/security/TLSConfigChangeEventConsumerTests.java
@@ -10,7 +10,7 @@
 package io.pravega.segmentstore.server.host.security;
 
 import io.netty.handler.ssl.SslContext;
-import io.pravega.test.common.SecurityConfigDefaults;
+import io.pravega.shared.security.utils.SecurityConfigDefaults;
 import org.junit.Test;
 import java.io.File;
 import java.nio.file.WatchEvent;
