+++ test/integration/src/main/java/io/pravega/test/integration/selftest/adapters/ExternalAdapter.java
@@ -15,7 +15,7 @@ import io.pravega.client.admin.StreamManager;
 import io.pravega.client.stream.StreamConfiguration;
 import io.pravega.client.stream.impl.DefaultCredentials;
 import io.pravega.common.Exceptions;
-import io.pravega.test.common.SecurityConfigDefaults;
+import io.pravega.shared.security.utils.SecurityConfigDefaults;
 import io.pravega.common.util.Retry;
 import io.pravega.test.integration.selftest.TestConfig;
 import java.net.URI;
